CONDEC-808: Remove calculation method. The Strategy Pattern does not really fit to the problem.
