+++ src/main/java/de/uhd/ifi/se/decision/management/jira/config/SettingsOfSingleProject.java
@@ -1,13 +1,6 @@
 package de.uhd.ifi.se.decision.management.jira.config;
 
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-import javax.inject.Inject;
-import javax.servlet.http.HttpServletRequest;
-
 import com.atlassian.templaterenderer.TemplateRenderer;
-
 import de.uhd.ifi.se.decision.management.jira.ComponentGetter;
 import de.uhd.ifi.se.decision.management.jira.classification.FileTrainer;
 import de.uhd.ifi.se.decision.management.jira.classification.implementation.OnlineFileTrainerImpl;
@@ -15,6 +8,11 @@ import de.uhd.ifi.se.decision.management.jira.classification.preprocessing.Prepr
 import de.uhd.ifi.se.decision.management.jira.model.DecisionKnowledgeProject;
 import de.uhd.ifi.se.decision.management.jira.persistence.ConfigPersistenceManager;
 import de.uhd.ifi.se.decision.management.jira.releasenotes.ReleaseNotesMapping;
+
+import javax.inject.Inject;
+import javax.servlet.http.HttpServletRequest;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * Renders the administration page to change the plug-in configuration of a
@@ -83,7 +81,6 @@ public class SettingsOfSingleProject extends AbstractSettingsServlet {
 		velocityParameters.put("projectSources",
 			ConfigPersistenceManager.getProjectSourcesForActiveProjects(projectKey));
 
-		velocityParameters.put("calculationMethod", ConfigPersistenceManager.getCalculationMethod(projectKey));
 		velocityParameters.put("addRecommendationDirectly", ConfigPersistenceManager.getAddRecommendationDirectly(projectKey));
 		velocityParameters.put("recommendationInput", ConfigPersistenceManager.getRecommendationInput(projectKey));
