--- src/main/java/de/uhd/ifi/se/decision/management/jira/decisionguidance/knowledgesources/calculationmethods/CalculationMethod.java
@@ -1,13 +1,0 @@
-package de.uhd.ifi.se.decision.management.jira.decisionguidance.knowledgesources.calculationmethods;
-
-import de.uhd.ifi.se.decision.management.jira.model.KnowledgeElement;
-import de.uhd.ifi.se.decision.management.jira.view.decisionguidance.Recommendation;
-
-import java.util.List;
-
-public interface CalculationMethod {
-
-	List<Recommendation> getResults(String inputs);
-	List<Recommendation> getResults(KnowledgeElement knowledgeElement);
-
-}
