--- src/main/java/de/uhd/ifi/se/decision/management/jira/decisionguidance/knowledgesources/calculationmethods/CalculationMethodType.java
@@ -1,6 +1,0 @@
-package de.uhd.ifi.se.decision.management.jira.decisionguidance.knowledgesources.calculationmethods;
-
-public enum CalculationMethodType {
-	SUBSTRING, TOKENIZED
-
-}
