+++ src/main/java/de/uhd/ifi/se/decision/management/jira/decisionguidance/recommender/IssueBasedRecommender.java
@@ -9,12 +9,14 @@ import java.util.List;
 
 public class IssueBasedRecommender extends BaseRecommender<KnowledgeElement> {
 
-	private KnowledgeElement knowledgeElement;
-
-	public IssueBasedRecommender(KnowledgeElement knowledgeElement) {
+	public IssueBasedRecommender() {
 		this.recommendations = new ArrayList<>();
 		this.knowledgeSources = new ArrayList<>();
-		this.knowledgeElement = knowledgeElement;
+	}
+
+	public IssueBasedRecommender(KnowledgeElement knowledgeElement) {
+		this();
+		this.input = knowledgeElement;
 	}
 
 	public IssueBasedRecommender(KnowledgeElement knowledgeElement, List<KnowledgeSource> knowledgeSources) {
@@ -24,8 +26,8 @@ public class IssueBasedRecommender extends BaseRecommender<KnowledgeElement> {
 
 	@Override
 	public List<Recommendation> getRecommendation() {
-		for (KnowledgeSource knowledgeSource : this.knowledgeSources) {
-			this.recommendations.addAll(knowledgeSource.getResults(this.knowledgeElement));
+		for (KnowledgeSource knowledgeSource : this.getActivatedKnowledgeSources()) {
+			this.recommendations.addAll(knowledgeSource.getResults(this.input));
 		}
 		return this.recommendations;
 	}
