+++ src/main/java/de/uhd/ifi/se/decision/management/jira/decisionguidance/recommender/RecommenderType.java
@@ -1,13 +1,13 @@
 package de.uhd.ifi.se.decision.management.jira.decisionguidance.recommender;
 
 public enum RecommenderType {
-	KEYWORD, ISSUE;
+	KEYWORD, ISSUE, EVALUATION;
 
 	/**
 	 * Change the default value here
 	 * @return
 	 */
-	public static String getDefault() {
-		return KEYWORD.toString();
+	public static RecommenderType getDefault() {
+		return KEYWORD;
 	}
 }
